requirements.txt
numpy
pandas
opencv-python
tqdm
torch
ultralytics
pillow
